doc_parser_fastapi/requirements.txt
fastapi
uvicorn
python-docx
python-docx2txt
pymupdf
pandas
requests
beautifulsoup4
lxml
striprtf
ebooklib
pyyaml
